+++ lib/beaker/dsl/install_utils.rb
@@ -1,10 +1,7 @@
-require 'beaker-puppet'
-
 module Beaker
   module DSL
     # Collection of installation methods and support
     module InstallUtils
-      include BeakerPuppet::InstallUtils
     end
   end
 end